restore secure user delete api method

+++ backend/routes/users.js
@@ -29,17 +29,21 @@ router.put("/:id", async (req, res) => {
 
 //DELETE
 router.delete("/:id", async (req, res) => {
-	try {
-		const user = await User.findById(req.params.id);
+	if (req.body.userId === req.params.id) {
 		try {
-			await Post.deleteMany({ username: user.username });
-			await User.findByIdAndDelete(req.params.id);
-			res.status(200).json("User has been deleted..");
+			const user = await User.findById(req.params.id);
+			try {
+				await Post.deleteMany({ username: user.username });
+				await User.findByIdAndDelete(req.params.id);
+				res.status(200).json("User has been deleted..");
+			} catch (err) {
+				res.status(500).json(err);
+			}
 		} catch (err) {
-			res.status(500).json(err);
+			res.status(404).json("User not found!");
 		}
-	} catch (err) {
-		res.status(404).json("User not found!");
+	} else {
+		res.status(401).json("You can delete only your account!");
 	}
 });
